Add agent, catalog, health, kv, and status endpoints
Renames consul-endpoints to be separated by files for easier file
management. Renames it to acl.go. This does require manual updates
to the generated go files to remove the duplicated Consul mock
object. This can be fixed in the next version of mock-http-api to
load a list of endpoint files for the same mock object.